+++ MelonLoader/Deprecated.cs
@@ -91,7 +91,6 @@ namespace MelonLoader
         public string Version { get; }
         public string Author { get; }
         public string DownloadLink { get; }
-
         public MelonModInfoAttribute(Type type, string name, string version, string author, string downloadLink = null)
         {
             SystemType = type;
@@ -124,7 +123,6 @@ namespace MelonLoader
         public string Version { get; }
         public string Author { get; }
         public string DownloadLink { get; }
-
         public MelonPluginInfoAttribute(Type type, string name, string version, string author, string downloadLink = null)
         {
             SystemType = type;
@@ -140,117 +138,49 @@ namespace MelonLoader
     [Obsolete("MelonPrefs is obsolete. Please use MelonPreferences instead.")]
     public class MelonPrefs
     {
-        private static string ConfigFileName = "modprefs.ini";
-        private static IniFile ConfigFile = null;
-        private static Dictionary<string, Dictionary<string, MelonPreference>> prefs = new Dictionary<string, Dictionary<string, MelonPreference>>();
-        private static Dictionary<string, string> categoryDisplayNames = new Dictionary<string, string>();
-        internal static void Setup() { if (ConfigFile == null) ConfigFile = new IniFile(Path.Combine(Core.UserDataPath, ConfigFileName)); }
-        public static void RegisterCategory(string name, string displayText) { categoryDisplayNames[name] = displayText; }
-        public static void RegisterString(string section, string name, string defaultValue, string displayText = null, bool hideFromList = false) { Register(section, name, defaultValue, displayText, MelonPreferenceType.STRING, hideFromList); }
-        public static void RegisterBool(string section, string name, bool defaultValue, string displayText = null, bool hideFromList = false) { Register(section, name, defaultValue ? "true" : "false", displayText, MelonPreferenceType.BOOL, hideFromList); }
-        public static void RegisterInt(string section, string name, int defaultValue, string displayText = null, bool hideFromList = false) { Register(section, name, "" + defaultValue, displayText, MelonPreferenceType.INT, hideFromList); }
-        public static void RegisterFloat(string section, string name, float defaultValue, string displayText = null, bool hideFromList = false) { Register(section, name, "" + defaultValue, displayText, MelonPreferenceType.FLOAT, hideFromList); }
-        private static void Register(string section, string name, string defaultValue, string displayText, MelonPreferenceType type, bool hideFromList)
+        public static void RegisterCategory(string name, string displayText) => MelonPreferences.CreateCategory(name, displayText);
+        public static void RegisterString(string section, string name, string defaultValue, string displayText = null, bool hideFromList = false) => MelonPreferences.GetCategory(name)?.CreateEntry(name, defaultValue, displayText, hideFromList);
+        public static void RegisterBool(string section, string name, bool defaultValue, string displayText = null, bool hideFromList = false) => MelonPreferences.GetCategory(name)?.CreateEntry(name, defaultValue, displayText, hideFromList);
+        public static void RegisterInt(string section, string name, int defaultValue, string displayText = null, bool hideFromList = false) => MelonPreferences.GetCategory(name)?.CreateEntry(name, defaultValue, displayText, hideFromList);
+        public static void RegisterFloat(string section, string name, float defaultValue, string displayText = null, bool hideFromList = false) => MelonPreferences.GetCategory(name)?.CreateEntry(name, defaultValue, displayText, hideFromList);
+        public static bool HasKey(string section, string name) => (MelonPreferences.GetCategory(section)?.GetEntry(name) != null);
+        public static Dictionary<string, Dictionary<string, MelonPreference>> GetPreferences()
         {
-            if (prefs.TryGetValue(section, out Dictionary<string, MelonPreference> prefsInSection))
+            Dictionary<string, Dictionary<string, MelonPreference>> output = new Dictionary<string, Dictionary<string, MelonPreference>>();
+            if (MelonPreferences.categorytbl.Count <= 0)
+                return output;
+            foreach (MelonPreferences_Category category in MelonPreferences.categorytbl)
             {
-                if (prefsInSection.TryGetValue(name, out MelonPreference pref))
-                    MelonLogger.LogError("Trying to registered Pref " + section + ":" + name + " more than one time");
-                else
+                Dictionary<string, MelonPreference> newprefsdict = new Dictionary<string, MelonPreference>();
+                foreach (MelonPreferences_Entry entry in category.prefstbl)
                 {
-                    string toStoreValue = defaultValue;
-                    if (ConfigFile.HasKey(section, name))
-                        toStoreValue = ConfigFile.GetString(section, name, defaultValue);
-                    else ConfigFile.SetString(section, name, defaultValue);
-                    prefsInSection.Add(name, new MelonPreference(toStoreValue, type, hideFromList, (displayText ?? "") == "" ? name : displayText));
+                    MelonPreference newpref = null;
+                    if (entry.Type == MelonPreferences_Entry.TypeEnum.STRING)
+                        newpref = new MelonPreference(entry.GetString(), (MelonPreferenceType)entry.Type, entry.Hidden, entry.DisplayName);
+                    else if (entry.Type == MelonPreferences_Entry.TypeEnum.BOOL)
+                        newpref = new MelonPreference(entry.GetBool().ToString(), (MelonPreferenceType)entry.Type, entry.Hidden, entry.DisplayName);
+                    else if (entry.Type == MelonPreferences_Entry.TypeEnum.INT)
+                        newpref = new MelonPreference(entry.GetInt().ToString(), (MelonPreferenceType)entry.Type, entry.Hidden, entry.DisplayName);
+                    else if (entry.Type == MelonPreferences_Entry.TypeEnum.FLOAT)
+                        newpref = new MelonPreference(entry.GetFloat().ToString(), (MelonPreferenceType)entry.Type, entry.Hidden, entry.DisplayName);
+                    if (newpref == null)
+                        continue;
+                    newprefsdict.Add(entry.Name, newpref);
                 }
+                output.Add(category.Name, newprefsdict);
             }
-            else
-            {
-                Dictionary<string, MelonPreference> dic = new Dictionary<string, MelonPreference>();
-                string toStoreValue = defaultValue;
-                if (ConfigFile.HasKey(section, name))
-                    toStoreValue = ConfigFile.GetString(section, name, defaultValue);
-                else ConfigFile.SetString(section, name, defaultValue);
-                dic.Add(name, new MelonPreference(toStoreValue, type, hideFromList, (displayText ?? "") == "" ? name : displayText));
-                prefs.Add(section, dic);
-            }
+            return output;
         }
-
-        public static bool HasKey(string section, string name) { return prefs.TryGetValue(section, out Dictionary<string, MelonPreference> prefsInSection) && prefsInSection.ContainsKey(name); }
-        public static Dictionary<string, Dictionary<string, MelonPreference>> GetPreferences() { return prefs; }
-        public static string GetCategoryDisplayName(string key) { if (categoryDisplayNames.TryGetValue(key, out string name)) return name; return key; }
-
-        public static void SaveConfig() => MelonHandler.SaveConfig();
-
-        internal static void SaveConfigToTable()
-        {
-            if (prefs.Count <= 0)
-                return;
-            bool has_saved = false;
-            foreach (KeyValuePair<string, Dictionary<string, MelonPreference>> prefsInSection in prefs)
-            {
-                if (prefsInSection.Value.Count <= 0)
-                    continue;
-                foreach (KeyValuePair<string, MelonPreference> pref in prefsInSection.Value)
-                {
-                    pref.Value.Value = pref.Value.ValueEdited;
-                    ConfigFile.SetString(prefsInSection.Key, pref.Key, pref.Value.Value);
-                    has_saved = true;
-                }
-            }
-            if (has_saved)
-                MelonLogger.Log("Legacy Config Saved!");
-        }
-
-        public static string GetString(string section, string name)
-        {
-            if (prefs.TryGetValue(section, out Dictionary<string, MelonPreference> prefsInSection) && prefsInSection.TryGetValue(name, out MelonPreference pref))
-                return pref.Value;
-            MelonLogger.LogError("Trying to get unregistered Pref " + section + ":" + name);
-            return "";
-        }
-
-        public static void SetString(string section, string name, string value)
-        {
-            if (prefs.TryGetValue(section, out Dictionary<string, MelonPreference> prefsInSection) && prefsInSection.TryGetValue(name, out MelonPreference pref))
-            {
-                pref.Value = pref.ValueEdited = value;
-                ConfigFile.SetString(section, name, value);
-            }
-            else
-                MelonLogger.LogError("Trying to save unknown pref " + section + ":" + name);
-        }
-
-        public static bool GetBool(string section, string name)
-        {
-            if (prefs.TryGetValue(section, out Dictionary<string, MelonPreference> prefsInSection) && prefsInSection.TryGetValue(name, out MelonPreference pref))
-                return (pref.Value.Equals("true") || pref.Value.Equals("1"));
-            MelonLogger.LogError("Trying to get unregistered Pref " + section + ":" + name);
-            return false;
-        }
-        public static void SetBool(string section, string name, bool value) { SetString(section, name, value ? "true" : "false"); }
-
-        public static int GetInt(string section, string name)
-        {
-            if (prefs.TryGetValue(section, out Dictionary<string, MelonPreference> prefsInSection) && prefsInSection.TryGetValue(name, out MelonPreference pref))
-                if (int.TryParse(pref.Value, out int valueI))
-                    return valueI;
-            MelonLogger.LogError("Trying to get unregistered Pref " + section + ":" + name);
-            return 0;
-        }
-        public static void SetInt(string section, string name, int value) { SetString(section, name, value.ToString()); }
-
-        public static float GetFloat(string section, string name)
-        {
-            if (prefs.TryGetValue(section, out Dictionary<string, MelonPreference> prefsInSection) && prefsInSection.TryGetValue(name, out MelonPreference pref))
-                if (float.TryParse(pref.Value, out float valueF))
-                    return valueF;
-            MelonLogger.LogError("Trying to get unregistered Pref " + section + ":" + name);
-            return 0.0f;
-        }
-        public static void SetFloat(string section, string name, float value) { SetString(section, name, value.ToString()); }
-
+        public static string GetCategoryDisplayName(string key) => MelonPreferences.GetCategory(key)?.DisplayName;
+        public static void SaveConfig() => MelonPreferences.Save();
+        public static string GetString(string section, string name) => MelonPreferences.GetCategory(section)?.GetEntry(name)?.GetString();
+        public static void SetString(string section, string name, string value) => MelonPreferences.GetCategory(section)?.GetEntry(name)?.SetString(value);
+        public static bool GetBool(string section, string name) => (bool)MelonPreferences.GetCategory(section)?.GetEntry(name)?.GetBool();
+        public static void SetBool(string section, string name, bool value) => MelonPreferences.GetCategory(section)?.GetEntry(name)?.SetBool(value);
+        public static int GetInt(string section, string name) => (int)MelonPreferences.GetCategory(section)?.GetEntry(name)?.GetInt();
+        public static void SetInt(string section, string name, int value) => MelonPreferences.GetCategory(section)?.GetEntry(name)?.SetInt(value);
+        public static float GetFloat(string section, string name) => (float)MelonPreferences.GetCategory(section)?.GetEntry(name)?.GetFloat();
+        public static void SetFloat(string section, string name, float value) => MelonPreferences.GetCategory(section)?.GetEntry(name)?.SetFloat(value);
         public enum MelonPreferenceType
         {
             STRING,
@@ -258,7 +188,6 @@ namespace MelonLoader
             INT,
             FLOAT
         }
-
         public class MelonPreference
         {
             public string Value { get; set; }