+++ MelonLoader/Melons/Preferences/API.cs
@@ -52,9 +52,7 @@ namespace MelonLoader
             }
         }
 
-        public static void Save() => MelonHandler.SaveConfig();
-
-        internal static void SaveConfig()
+        public static void Save()
         {
             if (categorytbl.Count <= 0)
                 return;
@@ -80,6 +78,7 @@ namespace MelonLoader
             }
             File.WriteAllText(FilePath, doc.ToString());
             MelonLogger.Msg("Config Saved!");
+            MelonHandler.OnPreferencesApplied();
         }
 
         public static MelonPreferences_Category GetCategory(string name)