+++ CHANGELOG.md
@@ -62,6 +62,8 @@
 42. Fixed Issue with Legacy MelonPrefs logging "Legacy Config Saved!" even when nothing happened.
 43. Fixed Issue with DEV Load Mode not working with ZIP Archives.
 44. Added Check to MelonHandler to make sure Plugins and Mods only load once from 1 DLL copy.
+45. Legacy MelonPrefs now properly redirect to MelonPreferences.
+46. 
 
 ---